refactor : hambuger_mapper -> hamburger_mapper로 변경
